+++ src/python/quad.h
@@ -1,0 +1,15 @@
+/*
+    quad.h -- Python bindings for numerical quadrature rules
+
+    Dr.Jit: A Just-In-Time-Compiler for Differentiable Rendering
+    Copyright 2025, Realistic Graphics Lab, EPFL.
+
+    All rights reserved. Use of this source code is governed by a
+    BSD-style license that can be found in the LICENSE.txt file.
+*/
+
+#pragma once
+
+#include "common.h"
+
+extern void export_quad(nb::module_ &m);